🧹 chore: satisfy swiftlint caps from #275 fix commits
Fallout from the two preceding fix commits:

- ModelDownloader.swift `download(...)` grew past `function_body_length`
  with the resumeData branching. Extract `mergeIntoDestination` private
  helper; the file-handling logic is the natural seam.
- ModelDownloader.swift class doc comment was orphaned by an inline
  rationale block. Fold the rationale into the doc comment as a
  `## Actor isolation` section.
- PromoCard.swift exceeded `file_length` (458 > 400) after the new ETA
  helpers + their `// Why:` block. Move the `extension PromoCard`
  helpers (computeSlotState, slotCopy, formatEta, computeEtaSeconds)
  to a sibling `PromoCard+Helpers.swift`. PromoCard.swift back to ~340
  lines.
- URLSessionModelDownloaderTests.swift `final class CapturingMockURLProtocol`
  tripped `static_over_final_class` on the `class func canInit/canonicalRequest`
  URLProtocol overrides. Drop `final` — the class is a test fixture and
  its non-finalness is harmless.

No behavior change.
